float16 to fix dataloading

+++ dexnet/torch_dataset.py
@@ -93,7 +93,7 @@ class Dex3Dataset(Dataset):
 
         def loader(idx):
             with np.load(key_to_npz(key, idx)) as x:
-                arr = x["arr_0"]
+                arr = x["arr_0"].astype(np.float16)
             return arr
 
         with ThreadPoolExecutor(max_workers=8) as executor:
